% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_naresponses.R
\name{clean_naresponses}
\alias{clean_naresponses}
\title{Clean na responses}
\usage{
clean_naresponses(word)
}
\arguments{
\item{word}{word to clean, if multiple use loop/apply/dplyr}
}
\description{
This function allows you to change responses that indicate lack of knowledge into NA.
}
\examples{
clean_naresponses('hello')
}
