% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_symbols.R
\name{clean_symbols}
\alias{clean_symbols}
\title{Remove symbols}
\usage{
clean_symbols(word)
}
\arguments{
\item{word}{word to clean, if multiple use loop/apply/dplyr}
}
\description{
This function allows you to remove symbols from responses.
}
\examples{
clean_symbols('hello')
}
