% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Full_Expand.R
\name{Full_Expand}
\alias{Full_Expand}
\title{Full Wordnet Expansion}
\usage{
Full_Expand(datax, antonym = T, syns = F)
}
\arguments{
\item{datax}{dataframe with three columns, named: terms, PoS, and sense.}

\item{antonym}{Obtain antonyms? Default is TRUE.}

\item{syns}{return synsets? Default is FALSE. IF FALSE, returns words.}
}
\description{
Gets see also, similar, attributes, hyponyms, antonyms, derivationally related forms, to expand a list of seed words.
}
\examples{
Full_Expand(data_with_columns_terms_PoS_&_sense)
}
